Projects/04_Agentic_Todo_Manager/llm-call/pyproject.toml
[project]
name = "llm-call"
version = "0.1.0"
description = "LLM backend for Agentic Todo Manager using OpenAI Agents SDK with Gemini via LiteLLM"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi[standard]",
    "uvicorn",
    "openai-agents[litellm]",
    "python-dotenv",
]
